Add Source Sans Pro to styles
Fix #460
The code to Source Sans Pro was removed at some point but it is still
defined as the main font for the body of the site.

Added Arial in replacement of Verdana for a fallback font since it is
proportionally very different from Source Sans Pro and the current type
scale makes Verdana seem very large compared to the headings.