Fix: Make 404 responses have the right status code
IIS changes the status code to 200 if a static page is served.
The only way I found to make it work is to "transform" the page
to an `ASP` one and set the status code at the beginning of it.

The main problem with this approach is that the content will not
be compressed, but that will be solved in a separate PR.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #483
Close #485