Fix: Verification with Bing and Google and update sitemap on push
After the migration from sonarwhal.com to webhint.io Google dropped
the verification. This adds the required files to make sure we are
verified in both search engines and adds an extra step in the build
process to tell them to fetch the sitemap so links in their cache
are updated faster.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #551
Close #555

Co-Authored-By: Cătălin Mariș <alrraa@gmail.com>