Fix: Enable `Clear-Site-Data` for sonarwhal.com
The old website had a service worker that it is still causing people to
see the old content instead of the new. One of the strategies to fix
this is to use the new `Clear-Site-Data` header supported by Chrome and
Firefox. This PR add this header header only for requests comming to
sonarwhal.com while leaving the rest untouched.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix: #544