Fix: Display full changelog
The project had different versions of the same package because of how
hoisting works, `hint`'s `optionalDependencies` (the configurations are
not the latest ones) and how npm handles duplicates and `package-lock.json`.
By removing `package-lock.json` npm will always go fetch the latest
version of the packages and avoid these issues. Also all packages will
be installed in `node_modules` and not inside another folder under it.

Because multiple versions of the same package, the changelog was not
always accurate.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #825